+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/completion/script/ParadoxScriptCompletionContributor.kt
@@ -32,7 +32,7 @@ class ParadoxScriptCompletionContributor : CompletionContributor() {
 
         // 当用户可能正在输入定值的命名空间或变量的名字时提示
         val defineNamePattern = psiElement()
-            .withElementType(ParadoxScriptTokenSets.KEY_TOKENS)
+            .withElementType(ParadoxScriptTokenSets.KEY_OR_STRING_TOKENS)
         extend(CompletionType.BASIC, defineNamePattern, ParadoxDefineNameCompletionProvider())
 
         // 当用户可能正在输入变量名时提示